o4/app-dev-2/projects/o2_quiz_ai/quizzes/serializers.py
from rest_framework import serializers
from .models import Quiz


class QuizSerializer(serializers.ModelSerializer):
    class Meta:
        model = Quiz
        fields = ["id", "title", "description", "created_at"]
