o4/app-dev-2/projects/o2_quiz_ai/requirements.txt
asgiref==3.9.2
Django==4.2.24
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
